+++ packages/create-sitecore-jss/src/templates/angular-xmcloud/src/app/routing/scripts/scripts.module.ts
@@ -1,0 +1,10 @@
+import { NgModule } from '@angular/core';
+import { ScriptsComponent } from './scripts.component';
+import { JssModule } from '@sitecore-jss/sitecore-jss-angular';
+
+@NgModule({
+  exports: [ScriptsComponent],
+  imports: [JssModule],
+  declarations: [ScriptsComponent],
+})
+export class ScriptsModule {}