+++ packages/sitecore-jss-angular/src/public_api.ts
@@ -10,6 +10,7 @@ export { RenderEmptyDirective } from './components/render-empty.directive';
 export { RenderComponentComponent } from './components/render-component.component';
 export { DateDirective } from './components/date.directive';
 export { EditFrameComponent } from './components/editframe.component';
+export { EditingScriptsComponent } from './components/editing-scripts.component';
 export {
   ComponentNameAndType,
   DYNAMIC_COMPONENT,